fixed soem env stuff

+++ k8s/scripts/init-env.sh
@@ -1,25 +1,28 @@
 
-echo "Port-forwarding into the cluster..."
-export PG_POD=$(kubectl get pods -n $NAMESPACE -l app=postgres -o custom-columns=:metadata.name --no-headers) # Get the name of the postgres pod
-echo "Postgres pod: $PG_POD" 
+if [ "$USING_DOCKER_COMPOSE" != "true" ]; then
+	echo "Port-forwarding into the cluster..."
+	export PG_POD=$(kubectl get pods -n $NAMESPACE -l app=postgres -o custom-columns=:metadata.name --no-headers) # Get the name of the postgres pod
+	echo "Postgres pod: $PG_POD" 
 
-TEMP_FILE=$(mktemp)
+	TEMP_FILE=$(mktemp)
 
-kubectl port-forward pod/$PG_POD 0:5432 -n $NAMESPACE > $TEMP_FILE 2>&1 &
+	kubectl port-forward pod/$PG_POD 0:5432 -n $NAMESPACE > $TEMP_FILE 2>&1 &
 
-PID=$! # Save the process ID to kill the port-forwarding later
-sleep 2 # Wait for the port-forwarding to start
-cat $TEMP_FILE
-LOCAL_PG_PORT=$(grep -o '127.0.0.1:[0-9]*' $TEMP_FILE | tail -n 1 | awk -F ':' '{print $2}')
+	PID=$! # Save the process ID to kill the port-forwarding later
+	sleep 2 # Wait for the port-forwarding to start
+	cat $TEMP_FILE
+	LOCAL_PG_PORT=$(grep -o '127.0.0.1:[0-9]*' $TEMP_FILE | tail -n 1 | awk -F ':' '{print $2}')
 
-trap "kill $PID" EXIT # Kill the port-forwarding when the script exits
+	trap "kill $PID" EXIT # Kill the port-forwarding when the script exits
 
+	echo "Port forwarding setup on local port: $LOCAL_PG_PORT"
+	rm $TEMP_FILE
 
-echo "Port forwarding setup on local port: $LOCAL_PG_PORT"
-rm $TEMP_FILE
-
-export PG_URL="postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@localhost:$LOCAL_PG_PORT/${POSTGRES_DB}"
-echo "PG_URL: $PG_URL"
+	export PG_URL="postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@localhost:$LOCAL_PG_PORT/${POSTGRES_DB}"
+	echo "PG_URL: $PG_URL"
+else
+	echo "Using Docker Compose. Skipping port-forwarding..."
+fi
